Implement title update for convo.
Expanded error handling in UserController, ConvoController, and MessageController to try/catch 500 errors when writing/deleting to the database.

Added messageCount to check if convo has any messages before deleting the messages.

+++ backend/controllers/ConvoController.js
@@ -108,12 +108,20 @@ exports.convo_create_post = asyncHandler(async (req, res, next) => {
     date_created: DateTime.now().toJSDate(),
   });
 
-  await newConvo.save();
+  try {
+    await newConvo.save();
 
-  return res.status(201).json({
-    success: true,
-    message: 'Convo created successfully',
-  });
+    return res.status(201).json({
+      success: true,
+      message: 'Convo created successfully',
+    });
+  } catch (error) {
+    console.error('Error:', error.message);
+    return res.status(500).json({
+      success: false,
+      message: 'Error creating convo',
+    });
+  }
 });
 
 // Handle convo delete on DELETE.
@@ -148,11 +156,15 @@ exports.convo_delete = asyncHandler(async (req, res, next) => {
   const session = await mongoose.startSession();
   session.startTransaction();
   try {
-    const deleteMessagesResult = await Message.deleteMany({convo: convo._id}).session(session).exec();
-    if (deleteMessagesResult.deletedCount === 0) {
-        throw new Error("Failed to delete messages");
-    } else {
-        console.log("Messages deleted");
+    const messageCount = await Message.countDocuments({convo: convo._id}).exec();
+    console.log(`There are ${messageCount} messages to delete`);
+    if (messageCount > 0) {
+      const deleteMessagesResult = await Message.deleteMany({convo: convo._id}).session(session).exec();
+      if (deleteMessagesResult.deletedCount === 0) {
+          throw new Error("Failed to delete messages");
+      } else {
+          console.log("Messages deleted");
+      }
     }
 
     const deletedConvo = await Convo.findByIdAndDelete(req.params.id).session(session).exec();
@@ -174,9 +186,6 @@ exports.convo_delete = asyncHandler(async (req, res, next) => {
       });
   }
   
-
-  
-
   return res.status(200).json({
     success: true,
     message: 'Convo deleted successfully',
@@ -185,7 +194,44 @@ exports.convo_delete = asyncHandler(async (req, res, next) => {
 
 // Handle convo update on PUT.
 exports.convo_title_update = asyncHandler(async (req, res, next) => {
-  res.send("NOT IMPLEMENTED: Convo update PUT");
+  // Parse the 'X-User' header to get the user object
+  const user = JSON.parse(req.headers['x-user']);
+  const userId = user.sub;
+  const mongoUser = await User.findOne({ auth0id: userId }).exec(); // MongoDB user
+
+  if (!mongoUser) {
+    return res.status(401).json({
+      success: false,
+      message: "Unauthorized",
+    });
+  }
+
+  const convo = await Convo.findById(req.params.id).exec();
+
+  if(!convo) {
+    return res.status(404).json({
+      success: false,
+      message: 'Convo not found',
+    });
+  }
+
+  let updatedTitle = req.body.title;
+  if (!updatedTitle.endsWith(" (edited)")) {
+      updatedTitle += " (edited)";
+  }
+  
+  const updatedConvo = await Convo.findByIdAndUpdate(req.params.id, { title: updatedTitle }).exec();
+  if(!updatedConvo) {
+    return res.status(500).json({
+      success: false,
+      message: 'Error updating convo title',
+    });
+  } else {
+    return res.status(201).json({
+      success: true,
+      message: 'Convo title updated successfully',
+    });
+  }
 });
 
 // Handle convo user add on PUT.
@@ -253,14 +299,22 @@ exports.convo_add_friend = asyncHandler(async (req, res, next) => {
     }
     
 
-    // If the friend is not already in the conversation, add them
-    convo.users.push(friend._id);
-    await convo.save();
+    try {
+      // If the friend is not already in the conversation, add them
+      convo.users.push(friend._id);
+      await convo.save();
 
-    return res.status(201).json({
-      success: true,
-      message: "Friend added to the conversation",
-    });
+      return res.status(201).json({
+        success: true,
+        message: "Friend added to the conversation",
+      });
+    } catch (error) {
+      console.error("Error adding friend to convo:", error);
+      return res.status(500).json({
+        success: false,
+        message: "Error adding friend to convo",
+      });
+    } 
 });
 
 // Handle convo user remove on PUT.
@@ -317,13 +371,22 @@ exports.convo_remove_friend = asyncHandler(async (req, res, next) => {
       });
     }
 
-    convo.users.pull(friend._id);
-    await convo.save();
+    try {
+      convo.users.pull(friend._id);
+      await convo.save();
 
-    return res.status(201).json({
-      success: true,
-      message: "Friend removed from the conversation",
-    });
+      return res.status(201).json({
+        success: true,
+        message: "Friend removed from the conversation",
+      });
+    } catch (error) {
+      console.error("Error removing friend from convo:", error);
+      return res.status(500).json({
+        success: false,
+        message: "Error removing friend from convo",
+      });
+    }
+    
 });
 
 exports.convo_leave = asyncHandler(async (req, res, next) => {
@@ -365,11 +428,20 @@ exports.convo_leave = asyncHandler(async (req, res, next) => {
     });
   }
 
-  convo.users.pull(mongoUser._id);
-  await convo.save();
 
-  return res.status(201).json({
-    success: true,
-    message: "You have left the conversation",
-  });
+  try {
+    convo.users.pull(mongoUser._id);
+    await convo.save();
+
+    return res.status(201).json({
+      success: true,
+      message: "You have left the conversation",
+    });
+  } catch (error) {
+    console.error("Error leaving convo:", error);
+    return res.status(500).json({
+      success: false,
+      message: "Error leaving convo",
+    });
+  }
 });
